// src/types/Athlete.tsx
export interface Athlete {
    name: string;
    cpf: string;
    rg: string;
    address: string;
    fatherName: string;
    motherName: string;
    birthDate: string;
    phoneNumber: string;
    password: string;
    email: string;
    responsibleName: string;
    responsibleEmail: string;
    motherPhoneNumber: string;
    fatherPhoneNumber: string;
    bloodType: string;
    frontIdPhotoUrl: string;
    backIdPhotoUrl: string;
    athletePhotoUrl: string;
    foodAllergies: string;
}